+++ library/NekMesh/Module/ProcessModules/ProcessRevolve.cpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-//  File: ProcessExtrude.cpp
+//  File: ProcessRevolve.cpp
 //
 //  For more information, please see: http://www.nektar.info/
 //
@@ -28,11 +28,12 @@
 //  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 //  DEALINGS IN THE SOFTWARE.
 //
-//  Description: Extrude a two-dimensional mesh to a three-dimensional mesh.
+//  Description: Revolve a two-dimensional mesh around the y-axis to form a
+//  three-dimensional mesh.  Arguments are "angle" and "layers"
 //
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "ProcessExtrude.h"
+#include "ProcessRevolve.h"
 #include <NekMesh/MeshElements/Element.h>
 
 using namespace std;
@@ -40,36 +41,40 @@ using namespace Nektar::NekMesh;
 
 namespace Nektar::NekMesh
 {
-ModuleKey ProcessExtrude::className =
+ModuleKey ProcessRevolve::className =
     GetModuleFactory().RegisterCreatorFunction(
-        ModuleKey(eProcessModule, "extrude"), ProcessExtrude::create);
+        ModuleKey(eProcessModule, "revolve"), ProcessRevolve::create);
 
-ProcessExtrude::ProcessExtrude(MeshSharedPtr m) : ProcessModule(m)
+ProcessRevolve::ProcessRevolve(MeshSharedPtr m) : ProcessModule(m)
 {
-    m_config["layers"] =
-        ConfigOption(false, "5", "Number of layers to extrude");
-    m_config["length"] = ConfigOption(false, "1.0", "Length of extrusion");
+    m_config["layers"] = ConfigOption(false, "24", "Number of layers");
+    m_config["angle"] =
+        ConfigOption(false, "full", "Angle of revolution (in radians)");
 }
 
-ProcessExtrude::~ProcessExtrude()
+ProcessRevolve::~ProcessRevolve()
 {
 }
 
-void ProcessExtrude::Process()
+void ProcessRevolve::Process()
 {
-    m_log(VERBOSE) << "Extruding grid." << endl;
+    m_log(VERBOSE) << "Revolving grid." << endl;
 
     if (m_mesh->m_spaceDim != 2)
     {
-        m_log(FATAL) << "Extrude should only be called for a two dimensional "
+        m_log(FATAL) << "Revolve should only be called for a two dimensional "
                      << "mesh" << endl;
     }
 
-    int nLayers      = m_config["layers"].as<int>();
-    NekDouble length = m_config["length"].as<NekDouble>();
-
-    NekDouble dz = length / nLayers;
-
+    int nLayers = m_config["layers"].as<int>();
+    // Special case needed for full revolutions
+    bool full       = m_config["angle"].as<std::string>() == "full";
+    NekDouble angle = full ? 2 * M_PI : m_config["angle"].as<NekDouble>();
+    if (angle < 0 || angle > 2 * M_PI)
+    {
+        m_log(FATAL) << "Angle should be between 0 and 2pi" << endl;
+    }
+    NekDouble dphi = angle / nLayers;
     // Increment space and expansion dimensions.
     m_mesh->m_spaceDim++;
     m_mesh->m_expDim++;
@@ -114,16 +119,30 @@ void ProcessExtrude::Process()
         id2node[n->m_id] = n;
     }
     // Save z plane coordinate
-    NekDouble z0 = 0;
+    NekDouble phi0 = 0;
 
     // Create vertices for subsequent layers.
     for (int i = 1; i < nLayers + 1; ++i)
     {
+        if (full && i == nLayers)
+        {
+            // For last layer we will connect up to first layer
+            break;
+        }
         for (auto &n : nodes)
         {
-            z0 = n->m_z;
-            NodeSharedPtr newNode(new Node(i * nodes.size() + n->m_id, n->m_x,
-                                           n->m_y, n->m_z + i * dz));
+            if (n->m_x <= 0)
+            {
+                // Disallow x=0 for now
+                m_log(FATAL)
+                    << "All vertices require positive (non-zero) x coordinates"
+                    << endl;
+            }
+            NekDouble r0 = sqrt(n->m_x * n->m_x + n->m_z * n->m_z);
+            NodeSharedPtr newNode(new Node(i * nodes.size() + n->m_id,
+                                           r0 * cos(i * dphi), n->m_y,
+                                           r0 * sin(i * dphi)));
+
             m_mesh->m_vertexSet.insert(newNode);
             id2node[i * nodes.size() + n->m_id] = newNode;
         }
@@ -133,20 +152,22 @@ void ProcessExtrude::Process()
 
     for (int j = 0; j < nLayers; ++j)
     {
+        int next = full ? (j + 1) % nLayers : j + 1;
         for (int i = 0; i < el.size(); ++i)
         {
             vector<NodeSharedPtr> verts = el[i]->GetVertexList();
             if (verts.size() == 4)
             {
                 vector<NodeSharedPtr> nodeList(8);
+
                 nodeList[0] = id2node[verts[0]->m_id + j * nodes.size()];
                 nodeList[1] = id2node[verts[1]->m_id + j * nodes.size()];
                 nodeList[2] = id2node[verts[2]->m_id + j * nodes.size()];
                 nodeList[3] = id2node[verts[3]->m_id + j * nodes.size()];
-                nodeList[4] = id2node[verts[0]->m_id + (j + 1) * nodes.size()];
-                nodeList[5] = id2node[verts[1]->m_id + (j + 1) * nodes.size()];
-                nodeList[6] = id2node[verts[2]->m_id + (j + 1) * nodes.size()];
-                nodeList[7] = id2node[verts[3]->m_id + (j + 1) * nodes.size()];
+                nodeList[4] = id2node[verts[0]->m_id + next * nodes.size()];
+                nodeList[5] = id2node[verts[1]->m_id + next * nodes.size()];
+                nodeList[6] = id2node[verts[2]->m_id + next * nodes.size()];
+                nodeList[7] = id2node[verts[3]->m_id + next * nodes.size()];
 
                 vector<int> tags(1);
                 tags[0] = 0;
@@ -161,11 +182,11 @@ void ProcessExtrude::Process()
             else
             {
                 vector<NodeSharedPtr> nodeList(6);
-                nodeList[0] = id2node[verts[0]->m_id + (j + 1) * nodes.size()];
-                nodeList[1] = id2node[verts[1]->m_id + (j + 1) * nodes.size()];
+                nodeList[0] = id2node[verts[0]->m_id + next * nodes.size()];
+                nodeList[1] = id2node[verts[1]->m_id + next * nodes.size()];
                 nodeList[2] = id2node[verts[1]->m_id + j * nodes.size()];
                 nodeList[3] = id2node[verts[0]->m_id + j * nodes.size()];
-                nodeList[4] = id2node[verts[2]->m_id + (j + 1) * nodes.size()];
+                nodeList[4] = id2node[verts[2]->m_id + next * nodes.size()];
                 nodeList[5] = id2node[verts[2]->m_id + j * nodes.size()];
 
                 vector<int> tags(1);
@@ -192,12 +213,19 @@ void ProcessExtrude::Process()
         {
             for (int j = 0; j < nLayers + 1; ++j)
             {
+                if (full && j == nLayers)
+                {
+                    // For last layer we will connect up to first layer
+                    break;
+                }
                 vector<NodeSharedPtr> ns(edge->m_edgeNodes.size());
                 for (int i = 0; i < ns.size(); i++)
                 {
                     NodeSharedPtr n = edge->m_edgeNodes[i];
-                    ns[i]           = std::shared_ptr<Node>(
-                        new Node(0, n->m_x, n->m_y, n->m_z + j * dz));
+                    NekDouble r0    = sqrt(n->m_x * n->m_x + n->m_z * n->m_z);
+
+                    ns[i] = std::shared_ptr<Node>(new Node(
+                        0, r0 * cos(j * dphi), n->m_y, r0 * sin(j * dphi)));
                 }
 
                 EdgeSharedPtr e = std::shared_ptr<Edge>(
@@ -344,12 +372,15 @@ void ProcessExtrude::Process()
         m_mesh->m_composite.insert(std::make_pair(itOc.second->m_id, comp));
     }
     // Create periodic composites
-    for (int i = 0; i < 2; i++)
+    if (!full)
     {
-        CompositeSharedPtr comp(new Composite());
-        comp->m_id  = ++maxCompId;
-        comp->m_tag = "F";
-        m_mesh->m_composite.insert(std::make_pair(maxCompId, comp));
+        for (int i = 0; i < 2; i++)
+        {
+            CompositeSharedPtr comp(new Composite());
+            comp->m_id  = ++maxCompId;
+            comp->m_tag = "F";
+            m_mesh->m_composite.insert(std::make_pair(maxCompId, comp));
+        }
     }
 
     // Populates boundary composites
@@ -365,9 +396,16 @@ void ProcessExtrude::Process()
                 {
                     for (int j = 0; j < 2; j++)
                     {
-                        if (LibUtilities::IsRealEqual(
-                                itQ->GetVertex(iV)->m_x,
-                                itOc.second->m_items[iEd]->GetVertex(j)->m_x) &&
+                        NekDouble oldr2 =
+                            itOc.second->m_items[iEd]->GetVertex(j)->m_x *
+                                itOc.second->m_items[iEd]->GetVertex(j)->m_x +
+                            itOc.second->m_items[iEd]->GetVertex(j)->m_z *
+                                itOc.second->m_items[iEd]->GetVertex(j)->m_z;
+                        NekDouble newr2 =
+                            itQ->GetVertex(iV)->m_x * itQ->GetVertex(iV)->m_x +
+                            itQ->GetVertex(iV)->m_z * itQ->GetVertex(iV)->m_z;
+
+                        if (LibUtilities::IsRealEqual(oldr2, newr2) &&
                             LibUtilities::IsRealEqual(
                                 itQ->GetVertex(iV)->m_y,
                                 itOc.second->m_items[iEd]->GetVertex(j)->m_y))
@@ -397,44 +435,50 @@ void ProcessExtrude::Process()
         }
 
         // Populates periodic composites
-        NekDouble zdist = 0.0;
-        for (int iV = 0; iV < itQ->GetVertexList().size(); iV++)
+        if (!full)
         {
-            zdist += itQ->GetVertex(iV)->m_z;
-        }
-        zdist                  = zdist / itQ->GetVertexList().size();
-        unsigned int compPerId = 0;
-        if (LibUtilities::IsRealEqual(zdist, z0))
-        {
-            compPerId = maxCompId - 1;
-        }
-        else if (LibUtilities::IsRealEqual(zdist - z0, length))
-        {
-            compPerId = maxCompId;
-        }
-        if (compPerId > 0 && itQ->GetVertexList().size() == 3)
-        {
-            // Triangle
-            auto perC = m_mesh->m_composite.find(compPerId);
-            ElmtConfig conf(LibUtilities::eTriangle, 1, false, false);
-            vector<int> tags(1);
-            tags[0]            = 1;
-            ElementSharedPtr E = GetElementFactory().CreateInstance(
-                LibUtilities::eTriangle, conf, itQ->GetVertexList(), tags);
-            E->SetId(itQ->GetId());
-            perC->second->m_items.push_back(E);
-        }
-        else if (compPerId > 0 && itQ->GetVertexList().size() == 4)
-        {
-            // Quad
-            auto perC = m_mesh->m_composite.find(compPerId);
-            ElmtConfig conf(LibUtilities::eQuadrilateral, 1, false, false);
-            vector<int> tags(1);
-            tags[0]            = 0;
-            ElementSharedPtr E = GetElementFactory().CreateInstance(
-                LibUtilities::eQuadrilateral, conf, itQ->GetVertexList(), tags);
-            E->SetId(itQ->GetId());
-            perC->second->m_items.push_back(E);
+            NekDouble phidist = 0.0;
+            for (int iV = 0; iV < itQ->GetVertexList().size(); iV++)
+            {
+                phidist +=
+                    (atan2(-itQ->GetVertex(iV)->m_z, -itQ->GetVertex(iV)->m_x) +
+                     M_PI);
+            }
+            phidist                = phidist / itQ->GetVertexList().size();
+            unsigned int compPerId = 0;
+            if (LibUtilities::IsRealEqual(phidist, phi0))
+            {
+                compPerId = maxCompId - 1;
+            }
+            else if (LibUtilities::IsRealEqual(phidist - phi0, angle))
+            {
+                compPerId = maxCompId;
+            }
+            if (compPerId > 0 && itQ->GetVertexList().size() == 3)
+            {
+                // Triangle
+                auto perC = m_mesh->m_composite.find(compPerId);
+                ElmtConfig conf(LibUtilities::eTriangle, 1, false, false);
+                vector<int> tags(1);
+                tags[0]            = 1;
+                ElementSharedPtr E = GetElementFactory().CreateInstance(
+                    LibUtilities::eTriangle, conf, itQ->GetVertexList(), tags);
+                E->SetId(itQ->GetId());
+                perC->second->m_items.push_back(E);
+            }
+            else if (compPerId > 0 && itQ->GetVertexList().size() == 4)
+            {
+                // Quad
+                auto perC = m_mesh->m_composite.find(compPerId);
+                ElmtConfig conf(LibUtilities::eQuadrilateral, 1, false, false);
+                vector<int> tags(1);
+                tags[0]            = 0;
+                ElementSharedPtr E = GetElementFactory().CreateInstance(
+                    LibUtilities::eQuadrilateral, conf, itQ->GetVertexList(),
+                    tags);
+                E->SetId(itQ->GetId());
+                perC->second->m_items.push_back(E);
+            }
         }
     }
     // Remove all faces composite